requirements.txt
Flask
google-generativeai
Pillow
pyttsx3
pyngrok
requests
